requirements.txt
python-dotenv
openai
langchain_community
langchain_openai
pyodbc
SQLAlchemy
requests